Use parser to control whitespace
