+++ lib/solid/parser/tag.ex
@@ -1,0 +1,3 @@
+defmodule Solid.Parser.Tag do
+  import NimbleParsec
+end
