+++ test/cases/112/input.liquid
@@ -1,4 +1,4 @@
 {% assign name = "Hans" %}
 Hello {{- name -}} !
-Hello {{ name -}} !
 Hello {{- name }} !
+Hello {{ name -}} !
